livekit-voice-agent/pyproject.toml
[project]
name = "livekit-voice-agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "googlemaps>=4.10.0",
    "livekit-agents[cartesia,deepgram,groq,openai,silero,turn-detector]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "pinecone>=7.3.0",
    "python-dotenv>=1.1.1",
]
